requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
cryptography==38.0.4
environs==9.5.0
future==0.18.2
marshmallow==3.19.0
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.21.0
python-telegram-bot==11.1.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
